backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
chromadb==0.4.18
google-generativeai==0.3.2
uuid==1.30
feedparser==6.0.10
python-multipart==0.0.6
starlette==0.27.0
jina>=3.15.1